requirements.txt
# Core dependencies for Gemini Voice Bot with RAG

# Google AI and Gemini
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pypdf>=3.17.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Vector storage and retrieval
chromadb>=0.4.18
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Audio processing
pyaudio>=0.2.13
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.3
colorama>=0.4.6

# Logging and monitoring
loguru>=0.7.2
